working on coupons

+++ database/migrations/2021_10_07_214446_create_customers_table.php
@@ -19,7 +19,7 @@ class CreateCustomersTable extends Migration
             $table->string('email')->nullable();
             $table->string('phone')->nullable();
             $table->string('address')->nullable();
-            $table->BigInteger('point')->nullable();
+            $table->integer('point')->nullable();
             $table->timestamps();
         });
     }
